ftdi-la: Implement acquisition using libusb
This fully removes our dependence on libftdi, clearing the way for us to
switch to asynchronous libusb calls and dramatically increase our
performance. As of this commit, though, we still use synchronous calls;
we just call them directly instead of via libftdi.

This commit isn't totally a no-op in terms of functionality, though. We
now have much better control over buffer sizes and the FTDI latency
timer, meaning samples arrive more frequently at lower baud rates.

Additionally, this commit fixes a bug in clock configuration that caused
sample rates on FT232R to be 113% too fast. (H-series chips were not
affected.) I assume this issue was never caught because clocking in
bitbang mode on the FT232R is unrecoverably broken regardless due to a
silicon bug. I plan to eventually drop support for the FT232R entirely
for this reason.