ftdi-la: Prevent dropped samples
The FTDI-LA driver currently drops samples periodically at any
moderately high sample rate. It has behaved this way since its
introduction, and its documentation explains this behavior as an
unavoidable limitation of the hardware, given its lack of overrun
reporting, its small on-chip buffers, and the fact it wasn't designed to
function as a logic analyzer.

That explanation is wrong. In reality, the driver drops samples because
it fails to poll the chip for new samples in a timely manner. Sample
acquisition currently happens via a blocking USB read called repeatedly
from a single thread. As a result, whenever that thread is descheduled
or busy with other work between calls, the USB bus is idle and the FTDI
chip is unable to empty its buffer. It doesn't take a very high sample
rate for this to cause device-side buffer overruns.

It may be tempting to blame this on the small buffer size of the FTDI
chips (4KB for H-series and 128B for the FT232R). However, the H-series
buffers aren't any smaller than those on FX2 logic analyzers--some of
the best-supported analyzers in sigrok today, which certainly never drop
samples. So while it's true that the FTDI chips don't seem to report
buffer overruns, a properly-written host application should be able to
stream data from them fast enough that it never becomes an issue to
begin with, at least for sample rates up to the order of 10MHz.

It turns out to be quite simple to make FTDI-LA behave properly: we can
ditch the blocking read calls in favor of libusb's asynchronous API,
which is designed for cases like this. This commit does exactly that.
Now, when we start acquisition, we queue up many read requests all at
once. libusb passes these requests to the kernel, allowing it to start
servicing each one immediately after its predecessor completes. As a
result, the USB bus is never idle and the chip can continuously empty
its buffers.

I copied a number of architectural choices and specific tunable values
from the fx2lafw driver, which uses the same libusb API in a very
similar manner.