ftdi-la: Drop support for FT232R
All silicon revisions of the FT232R have a silicon bug that results in a
huge amount of clock jitter in bitbang mode. As a result, they are not
suitable for use as logic analyzers. See FTDI TN_120 for more details on
the erratum.

Although FTDI-LA has nominally supported the FT232R since its
introduction, the erratum has always been present and has resulted in
data that cannot be parsed by any protocol decoder. In fact, until my
recent rewrite of the FTDI-LA protocol module, it configured an
incorrect clock rate (off by nearly 100%) for the FT232R. The fact that
no one reported or fixed this issue indicates just how broken the
clocking was to begin with.

Since this issue can't be worked around, drop support for the FT232R
entirely. To communicate this change to users, link to a new wiki entry
explaining the change when we detect an FT232R.