ftdi-la: Write sample rate to hardware at capture time
Currently, we write the sample rate to the hardware as soon as the
property gets set. This is inefficient, since the sample rate might be
changed many times before an actual capture starts. This commit defers
the actual hardware programming of sample rate to capture time but still
calculates the register values and does bounds checking at configuration
time.

This also lets us set up a default sample rate before the device is even
opened, which fixes a crash in PulseView when starting a capture before
explicitly setting a sample rate using the dropdown.